reorganized memo components
